refactor: relocate checkpoint quant metadata
