+++ python/sglang/srt/model_loader/weight_utils.py
@@ -40,7 +40,7 @@ from huggingface_hub import HfFileSystem, hf_hub_download, snapshot_download
 from pydantic import BaseModel, ConfigDict, ValidationInfo, model_validator
 from tqdm.auto import tqdm
 
-from sglang.checkpoint_quantization import resolve_checkpoint_quant_spec
+from sglang.srt.model_loader.checkpoint_quantization import resolve_checkpoint_quant_spec
 from sglang.srt.configs.load_config import LoadConfig
 from sglang.srt.configs.model_config import REQUANTIZATION_METHODS, ModelConfig
 from sglang.srt.distributed import (
